// lib/types.ts
export type Student = {
  id: string;
  census_record_1900: string;
  indian_name: string;
  family_name: string;
  english_given_name: string;
  alias: string;
  sex: string;
  year_of_birth: number;
  year_of_birth_uncertain: boolean;
  year_of_birth_uncertainty_type: string;
  year_of_birth_original_text: string;
  arrival_at_lincoln: Date;
  arrival_at_lincoln_uncertain: string;
  arrival_at_lincoln_uncertainty_type: string;
  arrival_at_lincoln_original_text: string;
  departure_from_lincoln: Date;
  departure_from_lincoln_uncertain: string;
  departure_from_lincoln_uncertainty_type: string;
  departure_from_lincoln_original_text: string;
  nation: string;
  band: string;
  agency: string;
  trade: string;
  source: string;
  comments: string;
  cause_of_death: string;
  cemetery_burial: string;
  relevant_links: string;
  [key: string]: unknown;
};